Moving the common helpers.
