+++ bitstring/bitarray_.py
@@ -10,7 +10,7 @@ from bitstring.exceptions import CreationError, Error
 from bitstring.bits import Bits, BitsType, TBits
 
 import bitstring.dtypes
-common_helpers = bitstring.bitstore_common_helpers
+helpers = bitstring.bitstore_helpers
 
 MutableBitStore = bitstring.bitstore.MutableBitStore
 
@@ -133,7 +133,7 @@ class BitArray(Bits):
     def fromstring(cls: TBits, s: str, /) -> TBits:
         """Create a new bitstring from a formatted string."""
         x = super().__new__(cls)
-        b = common_helpers.str_to_bitstore(s)
+        b = helpers.str_to_bitstore(s)
         x._bitstore = b._mutable_copy()
         return x
